+++ example/lib/reactive_form_builder_widget_sample.dart
@@ -23,7 +23,7 @@ class ReactiveFormBuilderWidgetSample extends StatelessWidget {
                     decoration: InputDecoration(
                       labelText: 'Name',
                     ),
-                    validationMessages: {
+                    validationMessages: (errors) => {
                       ValidationMessage.required: 'Name must not be empty',
                     },
                   ),