+++ lib/src/validators/credit_card_validator.dart
@@ -7,7 +7,7 @@ import 'package:reactive_forms/src/validators/number_validator.dart';
 
 /// A credit card validator that validates that the control's value is a valid
 /// credit card.
-class CreditCardValidator extends Validator {
+class CreditCardValidator extends Validator<dynamic> {
   @override
   Map<String, dynamic> validate(AbstractControl control) {
     final error = {ValidationMessage.creditCard: true};