+++ lib/src/value_accessors/int_value_accessor.dart
@@ -5,6 +5,8 @@
 import 'package:reactive_forms/reactive_forms.dart';
 import 'package:reactive_forms/src/value_accessors/control_value_accessor.dart';
 
+/// Represents a control value accessor that convert between data types
+/// [int] and [String].
 class IntValueAccessor extends ControlValueAccessor<int, String> {
   @override
   String modelToViewValue(int modelValue) {