+++ test/src/models/form_control_test.dart
@@ -43,7 +43,8 @@ void main() {
       );
 
       expect(formControl.errors.keys.length, 2);
-      expect(formControl.errors[ValidationMessage.email], true, reason: 'mail');
+      expect(formControl.errors.containsKey(ValidationMessage.email), true,
+          reason: 'mail');
       expect(formControl.errors[ValidationMessage.minLength] != null, true,
           reason: 'minLength');
     });
@@ -220,6 +221,55 @@ void main() {
 
       // Then: the control is pristine
       expect(control.pristine, true);
+    });
+
+    test('SetErrors marks control as dirty by default', () {
+      // Given: a control
+      final control = FormControl<String>();
+
+      // When: set errors
+      control.setErrors({'someError': true});
+
+      // Then: the control is dirty
+      expect(control.dirty, true);
+    });
+
+    test('Set errors and keep the control as pristine', () {
+      // Given: a control
+      final control = FormControl<String>();
+
+      // When: set errors and specify markAsDirty = false
+      control.setErrors({'someError': true}, markAsDirty: false);
+
+      // Then: the control is pristine
+      expect(control.pristine, true);
+    });
+
+    test('Checks if control has error', () {
+      // Given: a required and invalid control
+      final control = FormControl<String>(validators: [Validators.required]);
+
+      // Expect: control has Validators.required error
+      expect(control.hasError(ValidationMessage.required), true);
+    });
+
+    test('Checks if control has error', () {
+      // Given: a valid control
+      final control = FormControl<String>();
+
+      // Expect: control doesn't have Validators.required error
+      expect(control.hasError(ValidationMessage.required), false);
+    });
+
+    test('Reports control error', () {
+      // Given: a required and invalid control
+      final control = FormControl<String>(validators: [Validators.required]);
+
+      // When: get tha control error by errorCode
+      final error = control.getError(ValidationMessage.required);
+
+      // Then: the error is not null
+      expect(error, true);
     });
   });
 }