+++ test/src/widgets/reactive_form_builder_testing_widget.dart
@@ -3,7 +3,7 @@ import 'package:reactive_forms/reactive_forms.dart';
 
 class ReactiveFormBuilderTestingWidget extends StatelessWidget {
   final FormGroup form;
-  final Map<String, String> validationMessages;
+  final ValidationMessagesFunction validationMessages;
   final Map<String, String> bindings;
 
   const ReactiveFormBuilderTestingWidget({